requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==3.2.7
django-ckeditor==5.4.0
django-cleanup==2.1.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-extensions==3.1.5
django-js-asset==1.0.0
django-jsonfield==1.4.1
django-phonenumber-field==6.0.0
django-utils-six==2.0
phonenumbers==8.12.42
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.0.2
